// packages/fhir-r4/src/v4.0.1/capabilitystatementsecurity/types.ts
import type { Extension, Element, CodeableConcept } from "../core/types";

/** Generated from FHIR JSON Schema */

/** A Capability Statement documents a set of capabilities (behaviors) of a FHIR Server for a particular version of FHIR that may be used as a statement of actual server functionality or a statement of required or desired server implementation. */
export interface CapabilityStatementSecurity {
  id?: string;

  extension?: Extension[];

  modifierExtension?: Extension[];

  cors?: boolean;

  _cors?: Element;

  service?: CodeableConcept[];

  description?: string;

  _description?: Element;
}